♻️ refactor: list페이지에서 조공카드 매핑
